Resolve the final dependency cycle!
