+++ lib/codeql.js
@@ -23,14 +23,13 @@ var __importStar = (this && this.__importStar) || function (mod) {
     return result;
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.getTrapCachingExtractorConfigArgsForLang = exports.getTrapCachingExtractorConfigArgs = exports.getExtraOptions = exports.getCodeQLForCmd = exports.getCodeQLForTesting = exports.getCachedCodeQL = exports.setCodeQL = exports.getCodeQL = exports.setupCodeQL = exports.CODEQL_VERSION_RESOLVE_ENVIRONMENT = exports.CODEQL_VERSION_INIT_WITH_QLCONFIG = exports.CODEQL_VERSION_EXPORT_CODE_SCANNING_CONFIG = exports.CODEQL_VERSION_SECURITY_EXPERIMENTAL_SUITE = exports.CODEQL_VERSION_BETTER_RESOLVE_LANGUAGES = exports.CODEQL_VERSION_GHES_PACK_DOWNLOAD = exports.CommandInvocationError = void 0;
+exports.getGeneratedCodeScanningConfigPath = exports.getTrapCachingExtractorConfigArgsForLang = exports.getTrapCachingExtractorConfigArgs = exports.getExtraOptions = exports.getCodeQLForCmd = exports.getCodeQLForTesting = exports.getCachedCodeQL = exports.setCodeQL = exports.getCodeQL = exports.setupCodeQL = exports.CODEQL_VERSION_RESOLVE_ENVIRONMENT = exports.CODEQL_VERSION_INIT_WITH_QLCONFIG = exports.CODEQL_VERSION_EXPORT_CODE_SCANNING_CONFIG = exports.CODEQL_VERSION_SECURITY_EXPERIMENTAL_SUITE = exports.CODEQL_VERSION_BETTER_RESOLVE_LANGUAGES = exports.CODEQL_VERSION_GHES_PACK_DOWNLOAD = exports.CommandInvocationError = void 0;
 const fs = __importStar(require("fs"));
 const path = __importStar(require("path"));
 const core = __importStar(require("@actions/core"));
 const toolrunner = __importStar(require("@actions/exec/lib/toolrunner"));
 const yaml = __importStar(require("js-yaml"));
 const actions_util_1 = require("./actions-util");
-const config_utils_1 = require("./config-utils");
 const environment_1 = require("./environment");
 const error_matcher_1 = require("./error-matcher");
 const feature_flags_1 = require("./feature-flags");
@@ -734,7 +733,7 @@ async function generateCodeScanningConfig(codeql, config, features, logger) {
     if (!(await (0, feature_flags_1.useCodeScanningConfigInCli)(codeql, features))) {
         return;
     }
-    const codeScanningConfigFile = (0, config_utils_1.getGeneratedCodeScanningConfigPath)(config);
+    const codeScanningConfigFile = getGeneratedCodeScanningConfigPath(config);
     // make a copy so we can modify it
     const augmentedConfig = cloneObject(config.originalUserInput);
     // Inject the queries from the input
@@ -776,7 +775,7 @@ async function generateCodeScanningConfig(codeql, config, features, logger) {
     if (config.augmentationProperties.injectedMlQueries) {
         // We need to inject the ML queries into the original user input before
         // we pass this on to the CLI, to make sure these get run.
-        const packString = await (0, config_utils_1.getMlPoweredJsQueriesPack)(codeql);
+        const packString = await util.getMlPoweredJsQueriesPack(codeql);
         if (augmentedConfig.packs === undefined)
             augmentedConfig.packs = [];
         if (Array.isArray(augmentedConfig.packs)) {
@@ -805,7 +804,7 @@ function cloneObject(obj) {
  * Returns an empty list if a code scanning configuration file was not generated by the CLI.
  */
 async function getCodeScanningConfigExportArguments(config, codeql) {
-    const codeScanningConfigPath = (0, config_utils_1.getGeneratedCodeScanningConfigPath)(config);
+    const codeScanningConfigPath = getGeneratedCodeScanningConfigPath(config);
     if (fs.existsSync(codeScanningConfigPath) &&
         (await util.codeQlVersionAbove(codeql, exports.CODEQL_VERSION_EXPORT_CODE_SCANNING_CONFIG))) {
         return ["--sarif-codescanning-config", codeScanningConfigPath];
@@ -833,4 +832,13 @@ async function getTrapCachingExtractorConfigArgsForLang(config, language) {
     ];
 }
 exports.getTrapCachingExtractorConfigArgsForLang = getTrapCachingExtractorConfigArgsForLang;
+/**
+ * Get the path to the code scanning configuration generated by the CLI.
+ *
+ * This will not exist if the configuration is being parsed in the Action.
+ */
+function getGeneratedCodeScanningConfigPath(config) {
+    return path.resolve(config.tempDir, "user-config.yaml");
+}
+exports.getGeneratedCodeScanningConfigPath = getGeneratedCodeScanningConfigPath;
 //# sourceMappingURL=codeql.js.map
