+++ lib/config-utils.js
@@ -23,7 +23,7 @@ var __importStar = (this && this.__importStar) || function (mod) {
     return result;
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.getGeneratedCodeScanningConfigPath = exports.wrapEnvironment = exports.generateRegistries = exports.downloadPacks = exports.getConfig = exports.getPathToParsedConfigFile = exports.initConfig = exports.getMlPoweredJsQueriesStatus = exports.getMlPoweredJsQueriesPack = exports.ML_POWERED_JS_QUERIES_PACK_NAME = exports.parsePacks = exports.validatePackSpecification = exports.prettyPrintPack = exports.parsePacksSpecification = exports.parsePacksFromConfig = exports.calculateAugmentation = exports.getDefaultConfig = exports.getRawLanguages = exports.getLanguages = exports.getLanguagesInRepo = exports.getUnknownLanguagesError = exports.getNoLanguagesError = exports.getConfigFileDirectoryGivenMessage = exports.getConfigFileFormatInvalidMessage = exports.getConfigFileRepoFormatInvalidMessage = exports.getConfigFileDoesNotExistErrorMessage = exports.getConfigFileOutsideWorkspaceErrorMessage = exports.getLocalPathDoesNotExist = exports.getLocalPathOutsideOfRepository = exports.getPacksStrInvalid = exports.getPacksInvalid = exports.getPacksInvalidSplit = exports.getPathsInvalid = exports.getPathsIgnoreInvalid = exports.getQueryUsesInvalid = exports.getQueriesMissingUses = exports.getQueriesInvalid = exports.getDisableDefaultQueriesInvalid = exports.getNameInvalid = exports.validateAndSanitisePath = exports.defaultAugmentationProperties = void 0;
+exports.wrapEnvironment = exports.generateRegistries = exports.downloadPacks = exports.getConfig = exports.getPathToParsedConfigFile = exports.initConfig = exports.getMlPoweredJsQueriesStatus = exports.parsePacks = exports.validatePackSpecification = exports.parsePacksSpecification = exports.parsePacksFromConfig = exports.calculateAugmentation = exports.getDefaultConfig = exports.getRawLanguages = exports.getLanguages = exports.getLanguagesInRepo = exports.getUnknownLanguagesError = exports.getNoLanguagesError = exports.getConfigFileDirectoryGivenMessage = exports.getConfigFileFormatInvalidMessage = exports.getConfigFileRepoFormatInvalidMessage = exports.getConfigFileDoesNotExistErrorMessage = exports.getConfigFileOutsideWorkspaceErrorMessage = exports.getLocalPathDoesNotExist = exports.getLocalPathOutsideOfRepository = exports.getPacksStrInvalid = exports.getPacksInvalid = exports.getPacksInvalidSplit = exports.getPathsInvalid = exports.getPathsIgnoreInvalid = exports.getQueryUsesInvalid = exports.getQueriesMissingUses = exports.getQueriesInvalid = exports.getDisableDefaultQueriesInvalid = exports.getNameInvalid = exports.validateAndSanitisePath = exports.defaultAugmentationProperties = void 0;
 const fs = __importStar(require("fs"));
 const path = __importStar(require("path"));
 const perf_hooks_1 = require("perf_hooks");
@@ -165,7 +165,7 @@ async function addBuiltinSuiteQueries(languages, codeQL, resultMap, packs, suite
         if (!packs.javascript) {
             packs.javascript = [];
         }
-        packs.javascript.push(await getMlPoweredJsQueriesPack(codeQL));
+        packs.javascript.push(await (0, util_1.getMlPoweredJsQueriesPack)(codeQL));
         injectedMlQueries = true;
     }
     const suites = languages.map((l) => `${l}-${suiteName}.qls`);
@@ -173,7 +173,7 @@ async function addBuiltinSuiteQueries(languages, codeQL, resultMap, packs, suite
     return injectedMlQueries;
 }
 function isMlPoweredJsQueriesPack(pack) {
-    return parsePacksSpecification(pack).name === exports.ML_POWERED_JS_QUERIES_PACK_NAME;
+    return parsePacksSpecification(pack).name === util_1.ML_POWERED_JS_QUERIES_PACK_NAME;
 }
 /**
  * Retrieve the set of queries at localQueryPath and add them to resultMap.
@@ -870,12 +870,8 @@ function parsePacksSpecification(packStr, configFile) {
     };
 }
 exports.parsePacksSpecification = parsePacksSpecification;
-function prettyPrintPack(pack) {
-    return `${pack.name}${pack.version ? `@${pack.version}` : ""}${pack.path ? `:${pack.path}` : ""}`;
-}
-exports.prettyPrintPack = prettyPrintPack;
 function validatePackSpecification(pack, configFile) {
-    return prettyPrintPack(parsePacksSpecification(pack, configFile));
+    return (0, util_1.prettyPrintPack)(parsePacksSpecification(pack, configFile));
 }
 exports.validatePackSpecification = validatePackSpecification;
 // exported for testing
@@ -919,25 +915,6 @@ function combinePacks(packs1, packs2) {
     }
     return packs;
 }
-exports.ML_POWERED_JS_QUERIES_PACK_NAME = "codeql/javascript-experimental-atm-queries";
-/**
- * Gets the ML-powered JS query pack to add to the analysis if a repo is opted into the ML-powered
- * queries beta.
- */
-async function getMlPoweredJsQueriesPack(codeQL) {
-    let version;
-    if (await (0, util_1.codeQlVersionAbove)(codeQL, "2.11.3")) {
-        version = "~0.4.0";
-    }
-    else {
-        version = `~0.3.0`;
-    }
-    return prettyPrintPack({
-        name: exports.ML_POWERED_JS_QUERIES_PACK_NAME,
-        version,
-    });
-}
-exports.getMlPoweredJsQueriesPack = getMlPoweredJsQueriesPack;
 /**
  * Get information about ML-powered JS queries to populate status reports with.
  *
@@ -962,7 +939,7 @@ exports.getMlPoweredJsQueriesPack = getMlPoweredJsQueriesPack;
 function getMlPoweredJsQueriesStatus(config) {
     const mlPoweredJsQueryPacks = (config.packs.javascript || [])
         .map((p) => parsePacksSpecification(p))
-        .filter((pack) => pack.name === "codeql/javascript-experimental-atm-queries" && !pack.path);
+        .filter((pack) => pack.name === util_1.ML_POWERED_JS_QUERIES_PACK_NAME && !pack.path);
     switch (mlPoweredJsQueryPacks.length) {
         case 1:
             // We should always specify an explicit version string in `getMlPoweredJsQueriesPack`,
@@ -1241,13 +1218,4 @@ async function wrapEnvironment(env, operation) {
     }
 }
 exports.wrapEnvironment = wrapEnvironment;
-/**
- * Get the path to the code scanning configuration generated by the CLI.
- *
- * This will not exist if the configuration is being parsed in the Action.
- */
-function getGeneratedCodeScanningConfigPath(config) {
-    return path.resolve(config.tempDir, "user-config.yaml");
-}
-exports.getGeneratedCodeScanningConfigPath = getGeneratedCodeScanningConfigPath;
 //# sourceMappingURL=config-utils.js.map
