+++ lib/config-utils.test.js
@@ -35,7 +35,6 @@ const sinon = __importStar(require("sinon"));
 const api = __importStar(require("./api-client"));
 const codeql_1 = require("./codeql");
 const configUtils = __importStar(require("./config-utils"));
-const config_utils_1 = require("./config-utils");
 const feature_flags_1 = require("./feature-flags");
 const languages_1 = require("./languages");
 const logging_1 = require("./logging");
@@ -991,7 +990,7 @@ const packSpecPrettyPrintingMacro = ava_1.default.macro({
     exec: (t, packStr, packObj) => {
         const parsed = configUtils.parsePacksSpecification(packStr);
         t.deepEqual(parsed, packObj, "parsed pack spec is correct");
-        const stringified = configUtils.prettyPrintPack(packObj);
+        const stringified = (0, util_1.prettyPrintPack)(packObj);
         t.deepEqual(stringified, packStr.trim(), "pretty-printed pack spec is correct");
         t.deepEqual(configUtils.validatePackSpecification(packStr), packStr.trim(), "pack spec is valid");
     },
@@ -1463,23 +1462,23 @@ const ML_POWERED_JS_STATUS_TESTS = [
     // If another pack is loaded but not the ML-powered query pack, status is false.
     [["some-other/pack"], "false"],
     // If the ML-powered query pack is loaded with a specific version, status is that version.
-    [[`${config_utils_1.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.1.0`], "~0.1.0"],
+    [[`${util_1.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.1.0`], "~0.1.0"],
     // If the ML-powered query pack is loaded with a specific version and another pack is loaded, the
     // status is the version of the ML-powered query pack.
-    [["some-other/pack", `${config_utils_1.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.1.0`], "~0.1.0"],
+    [["some-other/pack", `${util_1.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.1.0`], "~0.1.0"],
     // If the ML-powered query pack is loaded without a version, the status is "latest".
-    [[config_utils_1.ML_POWERED_JS_QUERIES_PACK_NAME], "latest"],
+    [[util_1.ML_POWERED_JS_QUERIES_PACK_NAME], "latest"],
     // If the ML-powered query pack is loaded with two different versions, the status is "other".
     [
         [
-            `${config_utils_1.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.0.1`,
-            `${config_utils_1.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.0.2`,
+            `${util_1.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.0.1`,
+            `${util_1.ML_POWERED_JS_QUERIES_PACK_NAME}@~0.0.2`,
         ],
         "other",
     ],
     // If the ML-powered query pack is loaded with no specific version, and another pack is loaded,
     // the status is "latest".
-    [["some-other/pack", config_utils_1.ML_POWERED_JS_QUERIES_PACK_NAME], "latest"],
+    [["some-other/pack", util_1.ML_POWERED_JS_QUERIES_PACK_NAME], "latest"],
 ];
 for (const [packs, expectedStatus] of ML_POWERED_JS_STATUS_TESTS) {
     const packDescriptions = `[${packs
