+++ src/codeql.ts
@@ -7,11 +7,7 @@ import * as yaml from "js-yaml";
 
 import { getOptionalInput, isAnalyzingDefaultBranch } from "./actions-util";
 import * as api from "./api-client";
-import {
-  Config,
-  getGeneratedCodeScanningConfigPath,
-  getMlPoweredJsQueriesPack,
-} from "./config-utils";
+import type { Config } from "./config-utils";
 import { EnvVar } from "./environment";
 import { errorMatchers } from "./error-matcher";
 import {
@@ -1211,7 +1207,7 @@ async function generateCodeScanningConfig(
   if (config.augmentationProperties.injectedMlQueries) {
     // We need to inject the ML queries into the original user input before
     // we pass this on to the CLI, to make sure these get run.
-    const packString = await getMlPoweredJsQueriesPack(codeql);
+    const packString = await util.getMlPoweredJsQueriesPack(codeql);
 
     if (augmentedConfig.packs === undefined) augmentedConfig.packs = [];
     if (Array.isArray(augmentedConfig.packs)) {
@@ -1286,4 +1282,13 @@ export async function getTrapCachingExtractorConfigArgsForLang(
     `-O=${language}.trap.cache.bound=${TRAP_CACHE_SIZE_MB}`,
     `-O=${language}.trap.cache.write=${write}`,
   ];
+}
+
+/**
+ * Get the path to the code scanning configuration generated by the CLI.
+ *
+ * This will not exist if the configuration is being parsed in the Action.
+ */
+export function getGeneratedCodeScanningConfigPath(config: Config): string {
+  return path.resolve(config.tempDir, "user-config.yaml");
 }
