+++ src/config-utils.ts
@@ -28,7 +28,14 @@ import {
 import { Logger } from "./logging";
 import { RepositoryNwo } from "./repository";
 import { downloadTrapCaches } from "./trap-caching";
-import { codeQlVersionAbove, GitHubVersion, UserError } from "./util";
+import {
+  codeQlVersionAbove,
+  getMlPoweredJsQueriesPack,
+  GitHubVersion,
+  ML_POWERED_JS_QUERIES_PACK_NAME,
+  prettyPrintPack,
+  UserError,
+} from "./util";
 
 // Property names from the user-supplied config file.
 const NAME_PROPERTY = "name";
@@ -1592,12 +1599,6 @@ export function parsePacksSpecification(
   };
 }
 
-export function prettyPrintPack(pack: Pack) {
-  return `${pack.name}${pack.version ? `@${pack.version}` : ""}${
-    pack.path ? `:${pack.path}` : ""
-  }`;
-}
-
 export function validatePackSpecification(pack: string, configFile?: string) {
   return prettyPrintPack(parsePacksSpecification(pack, configFile));
 }
@@ -1663,28 +1664,6 @@ function combinePacks(packs1: Packs, packs2: Packs): Packs {
   return packs;
 }
 
-export const ML_POWERED_JS_QUERIES_PACK_NAME =
-  "codeql/javascript-experimental-atm-queries";
-
-/**
- * Gets the ML-powered JS query pack to add to the analysis if a repo is opted into the ML-powered
- * queries beta.
- */
-export async function getMlPoweredJsQueriesPack(
-  codeQL: CodeQL
-): Promise<string> {
-  let version;
-  if (await codeQlVersionAbove(codeQL, "2.11.3")) {
-    version = "~0.4.0";
-  } else {
-    version = `~0.3.0`;
-  }
-  return prettyPrintPack({
-    name: ML_POWERED_JS_QUERIES_PACK_NAME,
-    version,
-  });
-}
-
 /**
  * Get information about ML-powered JS queries to populate status reports with.
  *
@@ -1710,8 +1689,7 @@ export function getMlPoweredJsQueriesStatus(config: Config): string {
   const mlPoweredJsQueryPacks = (config.packs.javascript || [])
     .map((p) => parsePacksSpecification(p))
     .filter(
-      (pack) =>
-        pack.name === "codeql/javascript-experimental-atm-queries" && !pack.path
+      (pack) => pack.name === ML_POWERED_JS_QUERIES_PACK_NAME && !pack.path
     );
   switch (mlPoweredJsQueryPacks.length) {
     case 1:
@@ -2136,13 +2114,4 @@ export async function wrapEnvironment(
       process.env[key] = value;
     }
   }
-}
-
-/**
- * Get the path to the code scanning configuration generated by the CLI.
- *
- * This will not exist if the configuration is being parsed in the Action.
- */
-export function getGeneratedCodeScanningConfigPath(config: Config): string {
-  return path.resolve(config.tempDir, "user-config.yaml");
 }
